feat: port filter highlighting, line count badge & scroll-to-bottom to active LogsPage (closes #329)
Port the three useful features from PR #328 to the active LogsPage component:
- Filter match highlighting: Search term matches are highlighted in yellow within log lines
- Line count badge: Shows filtered/total line counts when searching in raw mode
- Scroll to Bottom button: ArrowDown icon button for manual jump-to-bottom when auto-refresh is paused

Also removes the orphaned pages/Logs.tsx file which was dead code with no imports.

Changes:
- Add highlightMatch() helper for regex-based search term highlighting with proper escaping
- Add allRawLines and filteredRawLines memos for efficient line splitting and filtering
- Replace parseAnsi-based pre rendering with individual line rendering for better highlighting support
- Add line count display in raw mode toolbar
- Add scroll-to-bottom button in raw mode toolbar with ArrowDown icon
- Import ArrowDown from lucide-react
